+++ src/app/inventory/item-types.ts
@@ -419,6 +419,9 @@ export interface DimPlugSet {
    * want to access DimSocket.emptyPlugItemHash instead!
    */
   readonly precomputedEmptyPlugItemHash?: number;
+
+  /** A precomputed list of plug hashes that can not roll on current versions of the item. */
+  readonly plugHashesThatCannotRoll: number[];
 }
 
 export interface DimSocket {
@@ -472,8 +475,6 @@ export interface DimSocket {
    */
   craftingData?: { [plugHash: number]: DestinyPlugItemCraftingRequirements | undefined };
 
-  /** Plug hashes in this item visible in the collections roll, if this is a perk */
-  curatedRoll: number[] | null;
   /**
    * The plug item hash used to reset this plug to an empty default plug.
    * This is a heuristic improvement over singleInitialItemHash, but it's