+++ src/app/inventory/store/sockets.ts
@@ -23,6 +23,7 @@ import {
 import deprecatedMods from 'data/d2/deprecated-mods.json';
 import { emptyPlugHashes } from 'data/d2/empty-plug-hashes';
 import {
+  BucketHashes,
   ItemCategoryHashes,
   PlugCategoryHashes,
   SocketCategoryHashes,
@@ -142,7 +143,10 @@ function buildInstancedSockets(
 
   return {
     allSockets: createdSockets, // Flat list of sockets
-    categories: categories.sort(compareBy((c) => c.category?.index)), // Sockets organized by category
+    categories:
+      itemDef.inventory?.bucketTypeHash === BucketHashes.Subclass
+        ? categories.sort(compareBy((c) => c.category?.index))
+        : categories, // Sockets organized by category
   };
 }
 
@@ -189,16 +193,20 @@ function buildDefinedSockets(
 
   return {
     allSockets: createdSockets, // Flat list of sockets
-    categories: categories.sort(compareBy((c) => c.category.index)), // Sockets organized by category
+    categories:
+      itemDef.inventory?.bucketTypeHash === BucketHashes.Subclass
+        ? categories.sort(compareBy((c) => c.category?.index))
+        : categories, // Sockets organized by category
   };
 }
 
 function filterReusablePlug(reusablePlug: DimPlug) {
-  const itemCategoryHashes = reusablePlug.plugDef.itemCategoryHashes || [];
   return (
-    !itemCategoryHashes.includes(ItemCategoryHashes.MasterworksMods) &&
-    !itemCategoryHashes.includes(ItemCategoryHashes.GhostModsProjections) &&
-    !reusablePlug.plugDef.plug?.plugCategoryIdentifier.includes('masterworks.stat')
+    !reusablePlug.plugDef.itemCategoryHashes?.some(
+      (ich) =>
+        ich === ItemCategoryHashes.MasterworksMods ||
+        ich === ItemCategoryHashes.GhostModsProjections
+    ) && !reusablePlug.plugDef.plug?.plugCategoryIdentifier.includes('masterworks.stat')
   );
 }
 
@@ -226,33 +234,35 @@ function buildDefinedSocket(
     return undefined;
   }
 
+  const isReusable = socketCategoryDef.categoryStyle === DestinySocketCategoryStyle.Reusable;
+  // This covers the visible intrinsic perks and armor stat plugs,
+  // for all of which everything but the currently plugged thing are distractions
+  const isIntrinsic = socketCategoryDef.categoryStyle === DestinySocketCategoryStyle.LargePerk;
+
   // Is this socket a perk-style socket, or something more general (mod-like)?
   const isPerk =
-    socketCategoryDef.categoryStyle === DestinySocketCategoryStyle.Reusable ||
     socketCategoryDef.categoryStyle === DestinySocketCategoryStyle.Unlockable ||
-    socketCategoryDef.categoryStyle === DestinySocketCategoryStyle.LargePerk;
-
-  const isReusable = socketCategoryDef.categoryStyle === DestinySocketCategoryStyle.Reusable;
+    isIntrinsic ||
+    isReusable;
 
   // The currently equipped plug, if any
   const reusablePlugs: DimPlug[] = [];
 
-  const craftingData: NonNullable<DimSocket['craftingData']> = {};
-
+  let craftingData: DimSocket['craftingData'];
   function addCraftingReqs(plugEntry: DestinyItemSocketEntryPlugItemRandomizedDefinition) {
     if (
       plugEntry.craftingRequirements &&
       (plugEntry.craftingRequirements.materialRequirementHashes.length ||
         plugEntry.craftingRequirements.unlockRequirements.length)
     ) {
-      craftingData[plugEntry.plugItemHash] = plugEntry.craftingRequirements;
+      (craftingData ??= {})[plugEntry.plugItemHash] = plugEntry.craftingRequirements;
     }
   }
 
   // We only build a larger list of plug options if this is a perk socket, since users would
   // only want to see (and search) the plug options for perks. For other socket types (mods, shaders, etc.)
   // we will only populate plugOptions with the currently inserted plug.
-  if (isPerk) {
+  if (isPerk && !isIntrinsic) {
     if (socketDef.reusablePlugSetHash) {
       const plugSet = defs.PlugSet.get(socketDef.reusablePlugSetHash, forThisItem);
       if (plugSet) {
@@ -342,11 +352,12 @@ function buildDefinedSocket(
   // if there's crafting data, sort plugs by their required level
   // TO-DO: the order is correct in the original plugset def,
   // we should address whatever is changing plug order in DIM
-  if (!_.isEmpty(craftingData)) {
+  if (craftingData) {
+    const cd = craftingData;
     plugOptions.sort(
-      compareBy((p) =>
+      compareBy((p: DimPlug) =>
         // shove retired perks to the bottom (our choice) and consider requiredLevel:undefined to be 0 (bungie data works this way)
-        p.cannotCurrentlyRoll ? 999 : craftingData[p.plugDef.hash]?.requiredLevel ?? 0
+        p.cannotCurrentlyRoll ? 999 : cd[p.plugDef.hash]?.requiredLevel ?? 0
       )
     );
   }
@@ -373,15 +384,14 @@ function buildDefinedSocket(
     plugged,
     plugOptions,
     plugSet,
-    curatedRoll: null,
     emptyPlugItemHash: findEmptyPlug(socketDef, socketTypeDef, plugSet),
-    reusablePlugItems: [],
+    reusablePlugItems: emptyArray(),
     hasRandomizedPlugItems:
       Boolean(socketDef.randomizedPlugSetHash) || socketTypeDef.alwaysRandomizeSockets,
     isPerk,
     isReusable,
     socketDefinition: socketDef,
-    craftingData: Object.keys(craftingData).length ? craftingData : undefined,
+    craftingData,
   };
 }
 
@@ -398,13 +408,12 @@ export function isPluggableItem(
 function isDestinyItemPlug(
   plug: DestinyItemPlugBase | DestinyItemSocketState
 ): plug is DestinyItemPlugBase {
-  return Boolean((plug as DestinyItemPlugBase).plugItemHash);
+  return 'plugItemHash' in plug;
 }
 
 function buildPlug(
   defs: D2ManifestDefinitions,
   plug: DestinyItemPlugBase | DestinyItemSocketState,
-  socketDef: DestinyItemSocketEntryDefinition,
   plugObjectivesData:
     | {
         [plugItemHash: number]: DestinyObjectiveProgress[];
@@ -412,41 +421,33 @@ function buildPlug(
     | undefined,
   plugSet: DimPlugSet | undefined
 ): DimPlug | null {
-  const plugHash = isDestinyItemPlug(plug) ? plug.plugItemHash : plug.plugHash;
-  const enabled = isDestinyItemPlug(plug) ? plug.enabled : plug.isEnabled;
+  const destinyItemPlug = isDestinyItemPlug(plug);
+  const plugHash = destinyItemPlug ? plug.plugItemHash : plug.plugHash;
 
   if (!plugHash) {
     return null;
   }
 
-  let plugDef = defs.InventoryItem.get(plugHash);
-  if (!plugDef && socketDef.singleInitialItemHash) {
-    plugDef = defs.InventoryItem.get(socketDef.singleInitialItemHash);
-  }
-
+  const plugDef = defs.InventoryItem.get(plugHash);
   if (!plugDef || !isPluggableItem(plugDef)) {
     return null;
   }
 
+  // These are almost never present
   const failReasons = plug.enableFailIndexes
     ? _.compact(
-        plug.enableFailIndexes.map((index) => plugDef.plug!.enabledRules[index]?.failureMessage)
+        plug.enableFailIndexes.map((index) => plugDef.plug.enabledRules[index]?.failureMessage)
       ).join('\n')
     : '';
 
-  const matchingPlugs =
-    plugSet?.plugs.filter((p) => p.plugDef.hash === plugDef.hash) ?? emptyArray();
-  const cannotCurrentlyRoll = Boolean(
-    matchingPlugs.length && matchingPlugs.every((p) => p.cannotCurrentlyRoll)
-  );
-
+  const enabled = destinyItemPlug ? plug.enabled : plug.isEnabled;
   return {
     plugDef,
-    enabled: enabled && (!isDestinyItemPlug(plug) || plug.canInsert),
+    enabled: enabled && (!destinyItemPlug || plug.canInsert),
     enableFailReasons: failReasons,
-    plugObjectives: plugObjectivesData?.[plugHash] || [],
+    plugObjectives: plugObjectivesData?.[plugHash] || emptyArray(),
     stats: null,
-    cannotCurrentlyRoll,
+    cannotCurrentlyRoll: plugSet?.plugHashesThatCannotRoll.includes(plugDef.hash),
   };
 }
 
@@ -460,32 +461,9 @@ export function buildDefinedPlug(defs: D2ManifestDefinitions, plugHash: number):
     plugDef,
     enabled: true,
     enableFailReasons: '',
-    plugObjectives: [],
+    plugObjectives: emptyArray(),
     stats: null,
   };
-}
-
-/**
- * A helper function to add plug options to a socket. This maintains the socketed plug's position in the list.
- */
-function addPlugOption(
-  built: DimPlug | null,
-  /** The active plug, which has already been built */
-  plug: DimPlug | null,
-  plugOptions: DimPlug[] // mutated
-) {
-  if (built && filterReusablePlug(built)) {
-    if (plug && built.plugDef.hash === plug.plugDef.hash) {
-      // Use the inserted plug we built earlier in this position, rather than the one we build from reusablePlugs.
-      plugOptions.shift();
-      plugOptions.push(plug);
-    } else {
-      // API Bugfix: Filter out intrinsic perks past the first: https://github.com/Bungie-net/api/issues/927
-      if (!built.plugDef.itemCategoryHashes?.includes(ItemCategoryHashes.WeaponModsIntrinsic)) {
-        plugOptions.push(built);
-      }
-    }
-  }
 }
 
 function isKnownEmptyPlugItemHash(plugItemHash: number) {
@@ -545,12 +523,14 @@ function findEmptyPlug(
   // leave emptyPlugItemHash set to undefined for sockets that never have
   // an empty plug, like abilities etc.
 
-  if (noDefaultSocketCategoryHashes.includes(socketType.socketCategoryHash)) {
-    return undefined;
-  }
-
-  // Y2+ weapon masterworks don't have an "empty" entry.
-  if (socket.socketTypeHash === weaponMasterworkY2SocketTypeHash) {
+  if (
+    // Sockets that ONLY get their items from your inventory necessarily can't be emptied
+    (socket.plugSources & ~SocketPlugSources.InventorySourced) === 0 ||
+    // Y2+ weapon masterworks don't have an "empty" entry.
+    socket.socketTypeHash === weaponMasterworkY2SocketTypeHash ||
+    // Socket categories that have no empty plug
+    noDefaultSocketCategoryHashes.includes(socketType.socketCategoryHash)
+  ) {
     return undefined;
   }
 
@@ -566,11 +546,6 @@ function findEmptyPlug(
     return undefined;
   }
 
-  // Sockets that ONLY get their items from your inventory necessarily can't be emptied
-  if ((socket.plugSources & ~SocketPlugSources.InventorySourced) === 0) {
-    return undefined;
-  }
-
   // Sometimes the empty plug is a regular plug set entry, sometimes it's one
   // of the reusablePlugItems. However, reusablePlugItems is thrown away when
   // there's a PlugSet, so we check the live API response reusablePlugs instead
@@ -579,9 +554,9 @@ function findEmptyPlug(
   // FIXME #7793: Retain socket.reusablePlugItems when it has unique items
   // and evaluate whether checking live API response is still necessary
   const empty =
-    reusablePlugs?.map((p) => p.plugItemHash).find(isKnownEmptyPlugItemHash) ||
+    reusablePlugs?.find((p) => isKnownEmptyPlugItemHash(p.plugItemHash))?.plugItemHash ||
     plugSet?.precomputedEmptyPlugItemHash ||
-    socket.reusablePlugItems.map((p) => p.plugItemHash).find(isKnownEmptyPlugItemHash);
+    socket.reusablePlugItems.find((p) => isKnownEmptyPlugItemHash(p.plugItemHash))?.plugItemHash;
 
   // Falling back to singleInitialItemHash is the conservative choice:
   // 1. Before this function existed, we used singleInitialItemHash all the
@@ -622,13 +597,16 @@ function buildSocket(
     return undefined;
   }
 
+  const isReusable = socketCategoryDef.categoryStyle === DestinySocketCategoryStyle.Reusable;
+  // This covers the visible intrinsic perks and armor stat plugs,
+  // for all of which everything but the currently plugged thing are distractions
+  const isIntrinsic = socketCategoryDef.categoryStyle === DestinySocketCategoryStyle.LargePerk;
+
   // Is this socket a perk-style socket, or something more general (mod-like)?
   const isPerk =
-    socketCategoryDef.categoryStyle === DestinySocketCategoryStyle.Reusable ||
     socketCategoryDef.categoryStyle === DestinySocketCategoryStyle.Unlockable ||
-    socketCategoryDef.categoryStyle === DestinySocketCategoryStyle.LargePerk;
-
-  const isReusable = socketCategoryDef.categoryStyle === DestinySocketCategoryStyle.Reusable;
+    isIntrinsic ||
+    isReusable;
 
   const plugSet = socketDef.reusablePlugSetHash
     ? buildCachedDimPlugSet(defs, socketDef.reusablePlugSetHash)
@@ -637,44 +615,67 @@ function buildSocket(
     : undefined;
 
   // The currently equipped plug, if any.
-  const plugged = buildPlug(defs, socket, socketDef, plugObjectivesData, plugSet);
-  // TODO: not sure if this should always be included!
-  const plugOptions = plugged ? [plugged] : [];
+  // This will always be one of the plugOptions -- either it's added
+  // as we look at all available plugs, or it's added after the loops.
+  const plugged = buildPlug(defs, socket, plugObjectivesData, plugSet);
+  let foundPluggedInOptions = false;
+  const plugOptions: DimPlug[] = [];
 
   // We only build a larger list of plug options if this is a perk socket, since users would
   // only want to see (and search) the plug options for perks. For other socket types (mods, shaders, etc.)
   // we will only populate plugOptions with the currently inserted plug.
-  let curatedRoll: number[] | null = null;
-  if (isPerk) {
+  if (isPerk && !isIntrinsic) {
     if (reusablePlugs) {
       // Get options from live info
       for (const reusablePlug of reusablePlugs) {
-        const built = buildPlug(defs, reusablePlug, socketDef, plugObjectivesData, plugSet);
-        addPlugOption(built, plugged, plugOptions);
+        if (plugged && reusablePlug.plugItemHash === plugged.plugDef.hash) {
+          plugOptions.push(plugged);
+          foundPluggedInOptions = true;
+        } else {
+          const built = buildPlug(defs, reusablePlug, plugObjectivesData, plugSet);
+          if (built && filterReusablePlug(built)) {
+            plugOptions.push(built);
+          }
+        }
       }
-      curatedRoll = socketDef.reusablePlugItems.map((p) => p.plugItemHash);
     } else if (socketDef.reusablePlugSetHash) {
       // Get options from plug set, instead of live info
       const plugSet = defs.PlugSet.get(socketDef.reusablePlugSetHash, forThisItem);
       if (plugSet) {
         for (const reusablePlug of plugSet.reusablePlugItems) {
-          const built = buildCachedDefinedPlug(
-            defs,
-            reusablePlug.plugItemHash,
-            reusablePlug.currentlyCanRoll
-          );
-          addPlugOption(built, plugged, plugOptions);
+          if (plugged && reusablePlug.plugItemHash === plugged.plugDef.hash) {
+            plugOptions.push(plugged);
+            foundPluggedInOptions = true;
+          } else {
+            const built = buildCachedDefinedPlug(
+              defs,
+              reusablePlug.plugItemHash,
+              reusablePlug.currentlyCanRoll
+            );
+            if (built && filterReusablePlug(built)) {
+              plugOptions.push(built);
+            }
+          }
         }
-        curatedRoll = plugSet.reusablePlugItems.map((p) => p.plugItemHash);
       }
     } else if (socketDef.reusablePlugItems) {
       // Get options from definition itself
       for (const reusablePlug of socketDef.reusablePlugItems) {
-        const built = buildCachedDefinedPlug(defs, reusablePlug.plugItemHash);
-        addPlugOption(built, plugged, plugOptions);
+        if (plugged && reusablePlug.plugItemHash === plugged.plugDef.hash) {
+          plugOptions.push(plugged);
+          foundPluggedInOptions = true;
+        } else {
+          const built = buildCachedDefinedPlug(defs, reusablePlug.plugItemHash);
+          if (built && filterReusablePlug(built)) {
+            plugOptions.push(built);
+          }
+        }
       }
-      curatedRoll = socketDef.reusablePlugItems.map((p) => p.plugItemHash);
     }
+  }
+
+  if (plugged && !foundPluggedInOptions) {
+    plugOptions.push(plugged);
   }
 
   // TODO: is this still true? also, should this be ?? instead of ||
@@ -686,7 +687,6 @@ function buildSocket(
     plugged,
     plugOptions,
     plugSet,
-    curatedRoll,
     emptyPlugItemHash: findEmptyPlug(socketDef, socketTypeDef, plugSet, reusablePlugs),
     hasRandomizedPlugItems,
     reusablePlugItems: reusablePlugs,
@@ -731,9 +731,12 @@ function buildCachedDimPlugSet(defs: D2ManifestDefinitions, plugSetHash: number)
   const dimPlugSet: DimPlugSet = {
     plugs,
     hash: plugSetHash,
-    precomputedEmptyPlugItemHash: defPlugSet.reusablePlugItems
-      .map((p) => p.plugItemHash)
-      .find(isKnownEmptyPlugItemHash),
+    precomputedEmptyPlugItemHash: defPlugSet.reusablePlugItems.find((p) =>
+      isKnownEmptyPlugItemHash(p.plugItemHash)
+    )?.plugItemHash,
+    plugHashesThatCannotRoll: plugs
+      .filter((p) => plugCannotCurrentlyRoll(plugs, p.plugDef.hash))
+      .map((p) => p.plugDef.hash),
   };
   reusablePlugSetCache[plugSetHash] = dimPlugSet;
 
@@ -762,4 +765,23 @@ function buildCachedDefinedPlug(
   definedPlugCache[plugHash] = plug;
 
   return plug ? (currentlyCanRoll === false ? { ...plug, cannotCurrentlyRoll: true } : plug) : null;
+}
+
+/**
+ * Determine if, given a plugSet, a given plug hash cannot roll. For this to be
+ * true, the plug hash must appear in the list of plugs in the plugSet, and all
+ * versions of that plug in the plugSet cannot currently roll.
+ */
+function plugCannotCurrentlyRoll(plugs: DimPlug[], plugHash: number) {
+  let matchingPlugs = false;
+  for (const p of plugs) {
+    if (p.plugDef.hash === plugHash) {
+      matchingPlugs = true;
+      if (!p.cannotCurrentlyRoll) {
+        return false; // we don't need to continue, we know it *can* roll
+      }
+    }
+  }
+  // There is at least one copy of the plug, and all matching copies cannot roll
+  return matchingPlugs;
 }