+++ src/app/search/search-filters/sockets.tsx
@@ -9,6 +9,7 @@ import {
   countEnhancedPerks,
   getIntrinsicArmorPerkSocket,
   getSocketsByCategoryHash,
+  matchesCuratedRoll,
 } from 'app/utils/socket-utils';
 import { StringLookup } from 'app/utils/util-types';
 import { DestinyItemSubType, DestinyRecordState } from 'bungie-api-ts/destiny2';
@@ -65,36 +66,20 @@ const socketFilters: FilterDefinition[] = [
     description: tl('Filter.RandomRoll'),
     destinyVersion: 2,
     filter: () => (item) =>
-      Boolean(item.energy) || item.sockets?.allSockets.some((s) => s.hasRandomizedPlugItems),
+      Boolean(item.bucket.inArmor && item.energy) ||
+      (!item.crafted &&
+        item.sockets?.allSockets.some(
+          (s) => s.isPerk && s.plugOptions.length > 0 && s.hasRandomizedPlugItems
+        )),
   },
   {
     keywords: 'curated',
     description: tl('Filter.Curated'),
     destinyVersion: 2,
-    filter: () => (item) => {
-      if (!item) {
-        return false;
-      }
-
-      const legendaryWeapon = item.bucket?.sort === 'Weapons' && item.tier === 'Legendary';
-
-      if (!legendaryWeapon) {
-        return false;
-      }
-
-      const matchesCollectionsRoll = item.sockets?.allSockets
-        // curatedRoll is only set for perk-style sockets
-        .filter((socket) => socket.plugOptions.length && socket.curatedRoll)
-        .every(
-          (socket) =>
-            socket.curatedRoll!.length === socket.plugOptions.length &&
-            socket.plugOptions.every(function (e, i) {
-              return e.plugDef.hash === socket.curatedRoll![i];
-            })
-        );
-
-      return matchesCollectionsRoll;
-    },
+    filter:
+      ({ d2Definitions }) =>
+      (item) =>
+        matchesCuratedRoll(d2Definitions!, item),
   },
   {
     keywords: 'extraperk',